Implemented the OpenChest Event
# Changelog

It's important to notice that there's deep stuff happening here, such as using a Co-Routine that has Actions that might have Actions as arguments. (wow there, bro). It's easy to understand once you read it carefully :smiley:

## Chest

- Activate invokes Game.Event.openChest
- Class methods such as
  - `changeToOpenSprite()`. Hardcodes the sprite to the openChestSprite
  - `spawnItem() `. This instantiates the getItem object.  The getItem object is the object that pops out from the chest and stays there until the DialogueBox finish.
  - `changeSprite()` Useful wrapper to just change a sprite without writing `spriteRenderer.sprite = <YOUR_SPRITE>`

## ImportChestTile

- A CustomTiledImporter to detect if a TileObject will add the Chest Component to the game object. It also hardcodes the sortingLayer since that's how our TiledMap works. This must be taken into account and it will probably has to be touched later.

## Audio

Added two sound effects for the Chest Script.

## Items Folder

This folder holds a SilverKey object that pops from the Chest. It also has the key's sprite. The Item that pops from the chest only



## Event

The most deeps things here is that we use Actions that receive Actions as arguments or stuff like that. Stuff like `Func<string[],Action,DialogueBox> createDialogueBox;` Which is a Func. that can optionally receive a lambda callback function.

### The OpenChest Event

This event needs many objects to fulfill it's functionality. It's very easy to read but there are some important things to notice.

- This event (and probably many others) will be IEnumerators, this helps us to use the WaitForSeconds function to delay other actions. In this event we wait for the sounds to play and the item to pop out of the chest and once that wait is done the DialogueBox will appear.
- There's still some hardcoded things but the Minimum functionality works and this will be tweaked slowly! :smile:

### The DisplayMessage Event

- Used for simple one-dialogue-only messages that are not on the conversations file.
- This event will receive a lambda function as a callback when the conversation ends.
- The reason it receives a lambda function it's because a displayMessageEnds (and once I tweak the startConversationEvent) it should be able to execute any encapsulated actions there. Eg. On this example when the chest opens and the dialogue ends, the popup Item is destroyed once the conversation ends.

## HUD

## createDialogueBox

Modified so it can receive an Action as a callback function to be executed once.



## DialogueBox

- Added attribute of endConversationCallback. This is passed to Game.Event.EndConversation